buildings/microservers/ai_service/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
httpx
pydantic
pydantic-settings
python-dotenv